Run podsandbox
